fix(new editor): fix PR after review
